requirements.txt
uvicorn==0.30.1
fastapi==0.111.1
fastapi-utils==0.7.0
grpcio==1.56.2
grpcio-tools==1.56.2
requests==2.32.3
redis==5.0.7
PyMySQL==1.1.1
SQLAlchemy==1.4.49
munch==4.0.0
typing_inspect==0.9.0
openai==1.37.1
python-consul==1.1.0
